OSIS-500 Move summary_responsible mgmt to attribution module
