+++ attribution/templates/summary_responsible.html
