+++ attribution/views/summary_responsible.py
@@ -1,0 +1,50 @@
+##############################################################################
+#
+#    OSIS stands for Open Student Information System. It's an application
+#    designed to manage the core business of higher education institutions,
+#    such as universities, faculties, institutes and professional schools.
+#    The core business involves the administration of students, teachers,
+#    courses, programs and so on.
+#
+#    Copyright (C) 2015-2017 Université catholique de Louvain (http://www.uclouvain.be)
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU General Public License as published by
+#    the Free Software Foundation, either version 3 of the License, or
+#    (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU General Public License for more details.
+#
+#    A copy of this license - GNU General Public License - is available
+#    at the root of the source code of this program.  If not,
+#    see http://www.gnu.org/licenses/.
+#
+##############################################################################
+from django.contrib.auth.decorators import login_required, user_passes_test
+from base import models as mdl_base
+from base.business.entity_version import find_entity_version_according_academic_year
+from base.models.entity_manager import is_entity_manager
+from base.views import layout
+
+
+@login_required
+@user_passes_test(is_entity_manager)
+def search_attributed_learning_units(request):
+    entities_manager = mdl_base.entity_manager.find_by_user(request.user)
+    academic_year = mdl_base.academic_year.current_academic_year()
+    _append_entity_version(entities_manager, academic_year)
+    return layout.render(request, 'search_attributed_learning_units.html', {"entities_manager": entities_manager,
+                                                              "academic_year": academic_year,
+                                                              "init": "0"})
+
+
+def _append_entity_version(entities_manager, academic_year):
+    for entity_manager in entities_manager:
+        if hasattr(entity_manager.entity, 'entity_versions') and entity_manager.entity.entity_versions:
+            entity_manager.entity_version = find_entity_version_according_academic_year(
+                entity_manager.entity.entity_versions, academic_year)
+        else:
+            entity_manager.entity_version = None
