+++ backoffice/urls.py
@@ -52,6 +52,8 @@ if 'dissertation' in settings.INSTALLED_APPS:
     urlpatterns += (url(r'^dissertation/', include('dissertation.urls')),)
 if 'assessments' in settings.INSTALLED_APPS:
     urlpatterns += (url(r'^assessments/', include('assessments.urls')),)
+if 'attribution' in settings.INSTALLED_APPS:
+    urlpatterns += (url(r'^attribution/', include('attribution.urls')),)
 
 handler404 = 'base.views.common.page_not_found'
 handler403 = 'base.views.common.access_denied'
